Fix deleteNode
What was wrong:
If `deleteNode()` was called on a node with children (e.g. delete parent[val=2] with children child[val=1] and child[val=3]) the node to be deleted gets replaced by one of its children. As a consequence the child has to be deleted since its value is being used in the parent. Deleting the child means freeing the variable and setting the parent-to-child pointer to NULL). However setting the parent's child pointer to NULL did not work.

How it has been fixed:
By reordering of replacement:
First: Get minimumSubNode/ parentNode and remove link to minimumSubNode
Second: deleteNode->val = maximumSubNode->val; // replace deleteNode with maximum of subtree

Why does that fix the issue?
Now the parentNode can correctly be identified and therefore the pointer to the child can be set to NULL correctly. Before, the BinarySearchTree was already modified with the child's value set into the parent, therefore not identifying the correct parent when calling getParentNode().